requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==41.0.5
pydantic==2.4.2
python-jose==3.3.0
APScheduler==3.10.4